+++ fs/verity/enable.c
@@ -215,7 +215,7 @@ static int enable_verity(struct file *filp,
 
 	/* Get the salt if the user provided one */
 	if (arg->salt_size &&
-	    copy_from_user(desc->salt, u64_to_user_ptr(arg->salt_ptr),
+	    copy_from_user(desc->salt, u64_to_user_ptr((u64)arg->salt_ptr),
 			   arg->salt_size)) {
 		err = -EFAULT;
 		goto out;
@@ -224,7 +224,7 @@ static int enable_verity(struct file *filp,
 
 	/* Get the signature if the user provided one */
 	if (arg->sig_size &&
-	    copy_from_user(desc->signature, u64_to_user_ptr(arg->sig_ptr),
+	    copy_from_user(desc->signature, u64_to_user_ptr((u64)arg->sig_ptr),
 			   arg->sig_size)) {
 		err = -EFAULT;
 		goto out;
